lets hope i dont break stuff next commit :D

+++ Proiect/Controllers/CategoriesController.cs
@@ -109,6 +109,12 @@ namespace Proiect.Controllers
             }
             else
             {
+                var hasProducts = db.Products.Any(p => p.CategoryId == category.Id);
+                if (hasProducts)
+                {
+                    TempData["message"] = "Categoria nu poate fi stearsa deoarece are produse asociate.";
+                    return RedirectToAction("Index");
+                }
                 db.Categories.Remove(category);
                 TempData["message"] = "Categoria a fost stearsa";
                 db.SaveChanges();
